// src/libs/types/editor-action.types.ts
export enum EditorActionType {
  ADD_ELEMENT = "ADD_ELEMENT",
  UPDATE_ELEMENT = "UPDATE_ELEMENT",
  DELETE_ELEMENT = "DELETE_ELEMENT",
  CHANGE_CLICKED_ELEMENT = "CHANGE_CLICKED_ELEMENT",
  CHANGE_DEVICE = "CHANGE_DEVICE",
  TOGGLE_PREVIEW_MODE = "TOGGLE_PREVIEW_MODE",
  REDO = "REDO",
  UNDO = "UNDO",
  LOAD_EDITOR = "LOAD_EDITOR",
}